+++ apps/api/src/components/components.controller.ts
@@ -14,6 +14,7 @@ import {
   ComponentDecision,
   ContentPartType,
   MessageRole,
+  ThreadMessage,
 } from '@use-hydra-ai/core';
 import {
   ChatMessage,
@@ -23,7 +24,7 @@ import {
 import { decryptProviderKey } from '../common/key.utils';
 import { CorrelationLoggerService } from '../common/services/logger.service';
 import { ProjectsService } from '../projects/projects.service';
-import { ThreadMessage } from '../threads/dto/message.dto';
+import { ThreadMessageDto } from '../threads/dto/message.dto';
 import { ThreadsService } from '../threads/threads.service';
 import {
   ComponentDecision as ComponentDecisionDto,
@@ -238,9 +239,21 @@ export class ComponentsController {
         true,
       );
 
-      //TODO: add 'in-progress' message to thread and update on each chunk
+      const tempId = new Date().toISOString();
+
       for await (const chunk of stream) {
-        response.write(`data: ${JSON.stringify(chunk)}\n\n`);
+        //TODO: don't create threadmessage here, add 'in-progress' message to thread and update on each chunk
+        const threadMessage: ThreadMessageDto = {
+          role: MessageRole.Hydra,
+          content: [{ type: ContentPartType.Text, text: chunk.message }],
+          id: tempId,
+          threadId: resolvedThreadId,
+          component: chunk,
+          createdAt: new Date(),
+          actionType: chunk.toolCallRequest ? ActionType.ToolCall : undefined,
+          toolCallRequest: chunk.toolCallRequest,
+        };
+        response.write(`data: ${JSON.stringify(threadMessage)}\n\n`);
       }
     } catch (error: any) {
       this.logger.error('Error in generateComponentStream:', error);
@@ -436,9 +449,20 @@ export class ComponentsController {
     );
 
     try {
-      //TODO: add 'in-progress' message to thread and update on each chunk
+      const tempId = new Date().toISOString();
       for await (const chunk of stream) {
-        response.write(`data: ${JSON.stringify(chunk)}\n\n`);
+        //TODO: don't create threadmessage here, add 'in-progress' message to thread and update on each chunk
+        const threadMessage: ThreadMessageDto = {
+          role: MessageRole.Hydra,
+          content: [{ type: ContentPartType.Text, text: chunk.message }],
+          id: tempId,
+          threadId: resolvedThreadId,
+          component: chunk,
+          createdAt: new Date(),
+          actionType: chunk.toolCallRequest ? ActionType.ToolCall : undefined,
+          toolCallRequest: chunk.toolCallRequest,
+        };
+        response.write(`data: ${JSON.stringify(threadMessage)}\n\n`);
       }
     } catch (error: any) {
       this.logger.error('Error in hydrateComponentStream:', error);
@@ -479,14 +503,23 @@ export class ComponentsController {
 }
 
 function convertThreadMessagesToLegacyThreadMessages(
-  currentThreadMessages: ThreadMessage[],
+  currentThreadMessages: ThreadMessage[] | ThreadMessageDto[],
 ) {
   return currentThreadMessages.map(
     (message): ChatMessage => ({
       sender: [MessageRole.User, MessageRole.Tool].includes(message.role)
         ? (message.role as 'user' | 'tool')
         : 'hydra',
-      message: message.content.map((part) => part.text ?? '').join(''),
+      message: message.content
+        .map((part) => {
+          switch (part.type) {
+            case ContentPartType.Text:
+              return part.text ?? '';
+            default:
+              return '';
+          }
+        })
+        .join(''),
     }),
   );
 }