+++ packages/hydra-ai-server/src/hydra-ai/hydra-ai-backend.ts
@@ -1,4 +1,4 @@
-import { ComponentDecision } from "@use-hydra-ai/core";
+import { ComponentDecision, ThreadMessage } from "@use-hydra-ai/core";
 import AIService from "./ai-service";
 import { ChatMessage } from "./model/chat-message";
 import {
@@ -6,8 +6,9 @@ import {
   AvailableComponents,
   ToolResponseBody,
 } from "./model/component-metadata";
-import { InputContext } from "./model/input-context";
+import { InputContext, InputContextAsArray } from "./model/input-context";
 import { Provider } from "./model/providers";
+import { SuggestionDecision } from "./services/suggestion/suggestion.types";
 
 interface HydraBackendOptions {
   version?: "v1" | "v2";
@@ -37,6 +38,41 @@ export default class HydraBackend {
     );
   }
 
+  public async generateSuggestions(
+    messageHistory: ThreadMessage[],
+    count: number,
+    availableComponents: AvailableComponent[],
+    threadId: string,
+    stream: true,
+  ): Promise<AsyncIterableIterator<SuggestionDecision>>;
+  public async generateSuggestions(
+    messageHistory: ThreadMessage[],
+    count: number,
+    availableComponents: AvailableComponent[],
+    threadId: string,
+    stream?: false | undefined,
+  ): Promise<SuggestionDecision>;
+  public async generateSuggestions(
+    messageHistory: ThreadMessage[],
+    count: number,
+    availableComponents: AvailableComponent[],
+    threadId: string,
+    stream?: boolean,
+  ): Promise<SuggestionDecision | AsyncIterableIterator<SuggestionDecision>> {
+    const context: InputContextAsArray = {
+      messageHistory: messageHistory ?? [],
+      availableComponents: availableComponents ?? [],
+      threadId,
+    };
+
+    return await this.aiService.generateSuggestions(
+      context,
+      count,
+      threadId,
+      stream,
+    );
+  }
+
   public async generateComponent(
     messageHistory: ChatMessage[],
     availableComponents: AvailableComponents,
@@ -58,7 +94,7 @@ export default class HydraBackend {
     const context: InputContext = {
       messageHistory,
       availableComponents,
-      threadId: threadId,
+      threadId,
     };
 
     return await this.aiService.chooseComponent(context, threadId, stream);
@@ -94,6 +130,7 @@ export default class HydraBackend {
     );
   }
 }
+
 /**
  * Generate a consistent, valid UUID from a string using SHA-256
  * This is used to ensure that the same string will always generate the same UUID