+++ packages/hydra-ai-server/src/hydra-ai/model/index.ts
@@ -2,4 +2,5 @@ export * from "./chat-message";
 export * from "./component-metadata";
 export * from "./component-props-metadata";
 export * from "./input-context";
+export * from "./openai-response";
 export * from "./providers";