+++ packages/hydra-ai-server/src/index.ts
@@ -3,3 +3,4 @@ export {
   default as HydraBackend,
 } from "./hydra-ai/hydra-ai-backend";
 export * from "./hydra-ai/model";
+export * from "./hydra-ai/services/suggestion/suggestion.types";